+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ v2.0.0
 Unreleased
 ------------------
 Added
- - `tcod.path.BasicGraph` for pathfinding on simple 2D arrays.
+ - `tcod.path.SimpleGraph` for pathfinding on simple 2D arrays.
 
 Changed
  - `tcod.path.CustomGraph` now accepts an `order` parameter.